Relay-Tracker/backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
atlassian-python-api==3.41.16
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
google-auth==2.37.0
libsql-experimental==0.0.49
sendgrid==6.11.0
